requirements.txt
SQLAlchemy==1.1.1
python-telegram-bot==10.1.0
requests==2.19.1
arrow==0.12.1
pycryptodome==3.6.4
emoji==0.4.5
pytils==0.3
google-cloud-vision==0.33.0
six==1.11.0
setuptools>=36.6.0
PyMySQL==0.9.2
redis==2.10.6
apiai==1.2.3
flask==1.0.2
flask-cors==3.0.6
Pillow==5.2.0
ImageHash==4.0
PySocks==1.6.8
